Monte carlo is finished

+++ core/src/com/group11/othello/AI/MonteCarlo/CarloNode.java
@@ -13,7 +13,7 @@ public class CarloNode {
     CarloNode parent;
     List<Vector2> moves;
     int playerTurn;
-    int visits;
+    double visits;
     double score;
     int row;
     int column;
@@ -24,8 +24,8 @@ public class CarloNode {
         moves = glCopy.getValidMoves(gl);
         playerTurn = glCopy.getTurnStatus();
         children = new ArrayList<CarloNode>();
-        visits = 0;
-        score = 0;
+        visits = 0.0;
+        score = 0.0;
         this.row = row;
         this.column = column;
 
@@ -50,7 +50,7 @@ public class CarloNode {
     }
     public double getScore(){return score;}
     public int getPlayerTurn() { return playerTurn; }
-    public int getVisits(){return visits;}
+    public double getVisits(){return visits;}
     public GameLogic getGameLogic(){return glCopy;}
     public  List<Vector2> getMoves(){return moves;}
     public int getRow(){return row;}
